+++ CHANGELOG.md
@@ -7,14 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 ### Added
+- The `characters_to_jyutping` function can now take a list of strings as input
+  with user-provided word segmentation.
+
 ### Changed
 - Updated versions of the dependencies: `pylangacq >= 0.17.0` and `wordseg >= 0.0.3`.
 - Restructured the repository to use top-level `src/` and `tests/` directories.
 
 ### Deprecated
 ### Removed
-- Removed `setup.py`.
-
 ### Fixed
 - Fixed known issues in the rime-cantonese data.